move connection ID generation from the utils to the protocol package
