+++ internal/protocol/connection_id.go
@@ -1,18 +1,19 @@
-package utils
+package protocol
 
 import (
 	"crypto/rand"
 	"encoding/binary"
-
-	"github.com/lucas-clemente/quic-go/internal/protocol"
 )
 
+// A ConnectionID in QUIC
+type ConnectionID uint64
+
 // GenerateConnectionID generates a connection ID using cryptographic random
-func GenerateConnectionID() (protocol.ConnectionID, error) {
+func GenerateConnectionID() (ConnectionID, error) {
 	b := make([]byte, 8)
 	_, err := rand.Read(b)
 	if err != nil {
 		return 0, err
 	}
-	return protocol.ConnectionID(binary.LittleEndian.Uint64(b)), nil
+	return ConnectionID(binary.LittleEndian.Uint64(b)), nil
 }
